chore: consolidate supabase helpers
